Move importer components to src/components/misc
